+++ routes/V1/admin/workers.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Modules\V1\AccessControl\Domain\ValueObjects\AdminPermissionEnum;
 use App\Modules\V1\Authentication\Domain\ValueObjects\TokenTypeEnum;
 use App\Modules\V1\Users\Domain\ValueObjects\PortalTypeEnum;
 use App\Modules\V1\Workers\Presentation\Http\Controllers\AdminWorkerController;
@@ -8,17 +9,17 @@ use Illuminate\Support\Facades\Route;
 Route::middleware('abilities:'.PortalTypeEnum::ADMIN->value.','.TokenTypeEnum::ACCESS_TOKEN->value)
     ->controller(AdminWorkerController::class)
     ->group(function () {
-        Route::get('/', 'index');
-        Route::post('/', 'store');
-        Route::post('/bulk-delete', 'bulkDelete');
+        Route::get('/', 'index')->middleware('permission:'.AdminPermissionEnum::VIEW_WORKER->value);
+        Route::post('/', 'store')->middleware('permission:'.AdminPermissionEnum::CREATE_WORKER->value);
+        Route::post('/bulk-delete', 'bulkDelete')->middleware('permission:'.AdminPermissionEnum::DELETE_WORKER->value);
         Route::prefix('trash')->group(function () {
-            Route::get('', 'trash');
-            Route::post('/bulk-restore', 'bulkRestore');
-            Route::delete('/bulk-force-delete', 'bulkForceDelete');
-            Route::post('/{id}/restore', 'restore');
-            Route::delete('/{id}', 'forceDelete');
+            Route::get('', 'trash')->middleware('permission:'.AdminPermissionEnum::VIEW_WORKER->value);
+            Route::post('/bulk-restore', 'bulkRestore')->middleware('permission:'.AdminPermissionEnum::EDIT_WORKER->value);
+            Route::delete('/bulk-force-delete', 'bulkForceDelete')->middleware('permission:'.AdminPermissionEnum::DELETE_WORKER->value);
+            Route::post('/{id}/restore', 'restore')->middleware('permission:'.AdminPermissionEnum::EDIT_WORKER->value);
+            Route::delete('/{id}', 'forceDelete')->middleware('permission:'.AdminPermissionEnum::DELETE_WORKER->value);
         });
-        Route::get('/{worker}', 'show');
-        Route::match(['put', 'patch'], '/{worker}', 'update');
-        Route::delete('/{worker}', 'destroy');
+        Route::get('/{worker}', 'show')->middleware('permission:'.AdminPermissionEnum::VIEW_WORKER->value);
+        Route::match(['put', 'patch'], '/{worker}', 'update')->middleware('permission:'.AdminPermissionEnum::EDIT_WORKER->value);
+        Route::delete('/{worker}', 'destroy')->middleware('permission:'.AdminPermissionEnum::DELETE_WORKER->value);
     });
